+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Removed dependencies on `http`, `bluebird`, `request`
 - Use customgrouping3 instead of customgrouping for categorization
 - Target `ES2022` instead of `ES5`
+- Various package versions
 
 ## Removed
 - Possibility to use a query string